add srs memory watcher.

+++ trunk/src/protocol/srs_protocol_buffer.cpp
@@ -72,7 +72,7 @@ SrsFastBuffer::~SrsFastBuffer()
 
 int SrsFastBuffer::size()
 {
-    return end - p;
+    return (int)(end - p);
 }
 
 char* SrsFastBuffer::bytes()
